EVG 3423 clean up comments and unnecessary code
