+++ operations/service.go
@@ -82,7 +82,7 @@ func startSystemCronJobs(ctx context.Context, env evergreen.Environment) {
 		units.PopulateTaskMonitoring(),
 		units.PopulateEventAlertProcessing(1),
 		units.PopulateBackgroundStatsJobs(env, 0),
-		units.PopulateLastContainerFinishTimeJobs(env)))
+		units.PopulateLastContainerFinishTimeJobs()))
 
 	amboy.IntervalQueueOperation(ctx, env.RemoteQueue(), 15*time.Second, time.Now(), opts, amboy.GroupQueueOperationFactory(
 		units.PopulateHostSetupJobs(env, 0),
